# openapi.yaml
openapi: 3.0.3
info:
  title: PAM Gateway API
  version: "1.0.0"
servers:
  - url: http://localhost:8080
security:
  - bearerAuth: []
paths:
  /api/v1/health:
    get:
      summary: Health check
      responses:
        "200":
          description: OK
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/targets:
    get:
      summary: List targets
      responses:
        "200":
          description: Targets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Target"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/targets/{id}:
    get:
      summary: Get target by id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Target
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Target"
        "404":
          $ref: "#/components/responses/NotFound"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/access/requests:
    post:
      summary: Create access request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AccessRequestCreate"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccessRequest"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "502":
          $ref: "#/components/responses/BadGateway"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/access/requests/{id}:
    get:
      summary: Get access request by id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Access request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccessRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/access/requests/{id}/approve:
    post:
      summary: Approve access request
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Approved request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccessRequest"
        "409":
          $ref: "#/components/responses/Conflict"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/sessions:
    post:
      summary: Create session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SessionCreate"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Session"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/sessions/{id}:
    get:
      summary: Get session by id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Session"
        "404":
          $ref: "#/components/responses/NotFound"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/sessions/{id}/terminate:
    post:
      summary: Terminate session
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Terminated session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Session"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/roles:
    get:
      summary: List roles
      responses:
        "200":
          description: Roles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Role"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"
    post:
      summary: Create role
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoleCreate"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Role"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/roles/{id}:
    get:
      summary: Get role by id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Role
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Role"
        "404":
          $ref: "#/components/responses/NotFound"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/policies:
    get:
      summary: List policies
      responses:
        "200":
          description: Policies
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Policy"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"
    post:
      summary: Create policy
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PolicyCreate"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Policy"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/policies/{id}:
    get:
      summary: Get policy by id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Policy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Policy"
        "404":
          $ref: "#/components/responses/NotFound"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/approvals:
    get:
      summary: List approvals
      responses:
        "200":
          description: Approvals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Approval"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"
    post:
      summary: Create approval
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApprovalCreate"
      responses:
        "200":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Approval"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/audit/events:
    get:
      summary: Query audit events
      parameters:
        - name: user
          in: query
          schema:
            type: string
        - name: target
          in: query
          schema:
            type: string
        - name: from
          in: query
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          schema:
            type: string
            format: date-time
      responses:
        "200":
          description: Audit events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AuditEvent"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/cmdb/sync:
    post:
      summary: Sync CMDB targets
      responses:
        "200":
          description: Sync result
          content:
            application/json:
              schema:
                type: object
                properties:
                  imported:
                    type: integer
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Conflict:
      description: Conflict
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    BadGateway:
      description: Bad gateway
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
        code:
          type: string
        details:
          type: object
          additionalProperties: true
    AccessRequestCreate:
      type: object
      required: [targetId, durationMinutes, reason]
      properties:
        targetId:
          type: string
        durationMinutes:
          type: integer
        reason:
          type: string
    AccessRequest:
      type: object
      properties:
        id:
          type: string
        targetId:
          type: string
        requestedBy:
          type: string
        durationMinutes:
          type: integer
        reason:
          type: string
        status:
          type: string
        createdAt:
          type: string
          format: date-time
        expiresAt:
          type: string
          format: date-time
        itsmKey:
          type: string
          nullable: true
    SessionCreate:
      type: object
      required: [targetId, requestId, protocol]
      properties:
        targetId:
          type: string
        requestId:
          type: string
        protocol:
          type: string
    Session:
      type: object
      properties:
        id:
          type: string
        targetId:
          type: string
        requestId:
          type: string
        protocol:
          type: string
        status:
          type: string
        startedAt:
          type: string
          format: date-time
        endedAt:
          type: string
          format: date-time
          nullable: true
    Target:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        type:
          type: string
        environment:
          type: string
        criticality:
          type: string
        status:
          type: string
    RoleCreate:
      type: object
      required: [name, description]
      properties:
        name:
          type: string
        description:
          type: string
    Role:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
    PolicyCreate:
      type: object
      required: [name, targetType, allowedProtocols, effect]
      properties:
        name:
          type: string
        targetType:
          type: string
        allowedProtocols:
          type: string
        effect:
          type: string
    Policy:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        targetType:
          type: string
        allowedProtocols:
          type: string
        effect:
          type: string
    ApprovalCreate:
      type: object
      required: [requestId, approver, status]
      properties:
        requestId:
          type: string
        approver:
          type: string
        status:
          type: string
    Approval:
      type: object
      properties:
        id:
          type: string
        requestId:
          type: string
        approver:
          type: string
        approvedAt:
          type: string
          format: date-time
        status:
          type: string
    AuditEvent:
      type: object
      properties:
        id:
          type: integer
        timestamp:
          type: string
          format: date-time
        eventType:
          type: string
        username:
          type: string
        role:
          type: string
        targetId:
          type: string
        targetName:
          type: string
        action:
          type: string
        result:
          type: string
        requestId:
          type: string
        sessionId:
          type: string
        sourceIp:
          type: string